+++ kiosk/app.py
@@ -2,6 +2,7 @@ from flask import Flask, render_template
 
 from kiosk.config import Config
 from kiosk.database import close_database, initialize_database
+from kiosk.features.api import api_blueprint
 
 
 def create_app():
@@ -9,6 +10,7 @@ def create_app():
     app.config.from_object(Config)
     app.teardown_appcontext(close_database)
     initialize_database(app)
+    app.register_blueprint(api_blueprint)
 
     @app.get("/kiosk")
     def kiosk_home():